+++ armory/data/adversarial/resisc45_densenet121_univpatch_and_univperturbation_adversarial_224x224.py
@@ -29,6 +29,7 @@ _CITATION = """
   publisher={IEEE}
 }
 """
+
 _URL = ""
 _DL_URL = (
     "https://armory-public-data.s3.us-east-2.amazonaws.com/adversarial-datasets/"
