+++ armory/data/adversarial_datasets.py
@@ -7,6 +7,7 @@ from typing import Callable
 from armory.data import datasets
 from armory.data.adversarial import imagenet_adversarial as IA  # noqa: F401
 from armory.data.adversarial import (  # noqa: F401
+    resisc45_densenet121_univpatch_and_univperturbation_adversarial_224x224,
     ucf101_mars_perturbation_and_patch_adversarial_112x112,
 )
 
@@ -40,6 +41,44 @@ def imagenet_adversarial(
         shuffle_files=False,
         cache_dataset=cache_dataset,
         framework=framework,
+    )
+
+
+def resisc45_adversarial_224x224(
+    split_type: str = "adversarial",
+    epochs: int = 1,
+    batch_size: int = 1,
+    dataset_dir: str = None,
+    preprocessing_fn: Callable = None,
+    cache_dataset: bool = False,
+    framework: str = "numpy",
+    clean_key: str = "clean",
+    adversarial_key: str = "adversarial_univpatch",
+) -> datasets.ArmoryDataGenerator:
+    """
+    resisc45 Adversarial Dataset of size (224, 224, 3),
+    including clean, adversarial universal perturbation, and
+    adversarial patched
+    """
+    if clean_key != "clean":
+        raise ValueError(f"{clean_key} != 'clean'")
+    adversarial_keys = ("adversarial_univpatch", "adversarial_univperturbation")
+    if adversarial_key not in adversarial_keys:
+        raise ValueError(f"{adversarial_key} not in {adversarial_keys}")
+
+    return datasets._generator_from_tfds(
+        "resisc45_densenet121_univpatch_and_univperturbation_adversarial224x224:1.0.0",
+        split_type=split_type,
+        batch_size=batch_size,
+        epochs=epochs,
+        dataset_dir=dataset_dir,
+        preprocessing_fn=preprocessing_fn,
+        as_supervised=False,
+        supervised_xy_keys=("images", "label"),
+        variable_length=False,
+        cache_dataset=cache_dataset,
+        framework=framework,
+        lambda_map=lambda x, y: ((x[clean_key], x[adversarial_key]), y),
     )
